requirements.txt
python-telegram-bot==20.6
google-cloud-speech
pydub
ffmpeg-python
tqdm
python-dotenv
